+++ aws/cloudwatch/src/main/scala-2.12/FutureWrapper.scala
@@ -1,0 +1,19 @@
+package com.nike.fleam.cloudwatch
+
+import scala.concurrent.Future
+import java.util.concurrent.{CompletionStage => JavaFuture}
+import scala.compat.java8.FutureConverters._
+
+/** Copyright 2020-present, Nike, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in
+ * the LICENSE file in the root directory of this source tree.
+ **/
+
+trait FutureWrapper {
+  // $COVERAGE-OFF$ Not worth testing
+  private[cloudwatch] def wrapRequest[Request, Response](f: Request => JavaFuture[Response]): Request => Future[Response] =
+    request => f(request).toScala
+  // $COVERAGE-ON$
+}