+++ aws/sqs/src/main/scala/com/nike/fleam/sqs/instances/ToMessageAttributes.scala
@@ -1,7 +1,7 @@
 package com.nike.fleam.sqs
 package instances
 
-import com.amazonaws.services.sqs.model.MessageAttributeValue
+import software.amazon.awssdk.services.sqs.model.MessageAttributeValue
 
 /** Copyright 2020-present, Nike, Inc.
  * All rights reserved.
@@ -14,19 +14,19 @@ trait ToMessageAttributesInstances {
   implicit val identityToMessageAttributes = ToMessageAttributes.lift[Map[String, MessageAttributeValue]](identity)
 
   implicit val tupleStringStringToMessageAttributes = ToMessageAttributes.lift[(String, String)] { case (key, value) =>
-    Map(key -> new MessageAttributeValue().withDataType("String").withStringValue(value))
+    Map(key -> MessageAttributeValue.builder().dataType("String").stringValue(value).build())
   }
 
   implicit val tupleStringIntToMessageAttributes = ToMessageAttributes.lift[(String, Int)] { case (key, value) =>
-    Map(key -> new MessageAttributeValue().withDataType("Number").withStringValue(value.toString))
+    Map(key -> MessageAttributeValue.builder().dataType("Number").stringValue(value.toString).build())
   }
 
   implicit val tupleStringLongToMessageAttributes = ToMessageAttributes.lift[(String, Long)] { case (key, value) =>
-    Map(key -> new MessageAttributeValue().withDataType("Number").withStringValue(value.toString))
+    Map(key -> MessageAttributeValue.builder().dataType("Number").stringValue(value.toString).build())
   }
 
   implicit val tupleStringDoubleToMessageAttributes = ToMessageAttributes.lift[(String, Double)] { case (key, value) =>
-    Map(key -> new MessageAttributeValue().withDataType("Number").withStringValue(value.toString))
+    Map(key -> MessageAttributeValue.builder().dataType("Number").stringValue(value.toString).build())
   }
 
   implicit def tupleMapStringTToMessageAttributes[T](implicit single: ToMessageAttributes[(String, T)]) =